backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Database ──────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0

# ── Auth ──────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# ── Validation ────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ── Machine Learning (Python 3.14 compatible versions) ────
scikit-learn==1.6.1
pandas==2.2.2
numpy==2.2.4
joblib==1.4.2

# ── Utilities ─────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.0

# ── Testing ───────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
aiosqlite==0.20.0